+++ CHANGELOG.md
@@ -17,11 +17,14 @@ Please mark all change in change log and use the issue from GitHub
 -   \#1661 HNSW support deleted vectors searching
 -   \#1825 Add annoy index type in C++ sdk
 -   \#1849 NSG support deleted vectors searching
+-   \#1893 Log config information and device information
 
 ## Improvement
 -   \#1627 Move read/write index APIs into codec
 -   \#1784 Add Substructure and Superstructure in http module
 -   \#1858 Disable S3 build
+-   \#1882 Add index annoy into http module
+-   \#1886 Refactor log on search and insert request
 
 ## Task